requirements.txt
dj_rest_auth==4.0.1
Django==4.2.3
djangorestframework==3.14.0
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.33.0
cryptography==41.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-shortcuts==1.6
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pip==23.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
